pyproject.toml
[tool.poetry]
name = "community-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Andrii Kudenko <andrii.kudenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "3.7.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.14.0"
asyncpg = "^0.30.0"
syncpg = "^1.1.3"
python-dotenv = "^1.0.1"
googlemaps = "^4.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
